Moved GetPrefix() out of common and into server/config/consul
